Add style for title and address
