+++ src/views/CoreUnitBudgetStatement/transparencyReportUtils.tsx
@@ -10,12 +10,12 @@ import { getPageWrapper } from '@ses/core/utils/dom';
 import lightTheme from '@ses/styles/theme/themes';
 import MobileDetect from 'mobile-detect';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
+import { WalletTableCell } from '@/components/WalletTableCell/WalletTableCell';
 import Information from '@/components/icons/information';
 import ArrowPopoverTargetValueComponent from '@/views/CoreUnitBudgetStatement/components/ArrowPopoverTargetValue/ArrowPopoverTargetValueComponent';
 import { formatAddressForOutput } from '../../core/utils/string';
 import { CustomLink } from '../../stories/components/CustomLink/CustomLink';
 import { TextCell } from '../../stories/components/TextCell/TextCell';
-import { WalletTableCell } from '../../stories/components/WalletTableCell/WalletTableCell';
 import ModalSheetValueContent from './components/TransparencyTransferRequest/components/ModalSheet/ModalSheetValueContent';
 import type { BudgetStatementWalletDto } from '../../core/models/dto/coreUnitDTO';
 import type { TargetBalanceTooltipInformation } from '@ses/core/utils/typesHelpers';
